# lib/fs: Add ASCII fastpath for normalization