### Purpose

Avoids `unicode.ToLower(unicode.ToUpper(r))` and `norm.NFC.String(s)` for simple ASCII filenames.