@calmh 2024-01-23
Test case and benchmark... how much faster is this than the old stuff, for the various cases, since it's a performance optimisation?