@bt90 2024-01-24
Added a benchmark. Comparison with the previous implementation:

```
goos: linux
goarch: amd64
pkg: github.com/syncthing/syncthing/lib/fs
cpu: AMD EPYC 7763 64-Core Processor                
                                     │ /tmp/old.txt  │            /tmp/new.txt             │
                                     │    sec/op     │   sec/op     vs base                │
UnicodeLowercase/lowercase_ASCII-2      48.34n ±  7%   17.25n ± 8%  -64.32% (p=0.000 n=10)
UnicodeLowercase/mixedcase_ASCII-2     191.45n ± 14%   75.91n ± 6%  -60.35% (p=0.000 n=10)
UnicodeLowercase/lowercase_unicode-2    127.2n ±  4%   235.0n ± 7%  +84.75% (p=0.000 n=10)
UnicodeLowercase/mixedcase_unicode-2    347.2n ±  7%   315.6n ± 6%   -9.13% (p=0.001 n=10)
UnicodeLowercase/multibyte_unicode-2    537.2n ±  6%   586.1n ± 3%   +9.10% (p=0.000 n=10)
geomean                                 185.5n         141.6n       -23.67%

                                     │ /tmp/old.txt │            /tmp/new.txt             │
                                     │     B/op     │    B/op     vs base                 │
UnicodeLowercase/lowercase_ASCII-2     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/mixedcase_ASCII-2     24.00 ± 0%     24.00 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/lowercase_unicode-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/mixedcase_unicode-2   32.00 ± 0%     32.00 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/multibyte_unicode-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                           ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                     │ /tmp/old.txt │            /tmp/new.txt             │
                                     │  allocs/op   │ allocs/op   vs base                 │
UnicodeLowercase/lowercase_ASCII-2     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/mixedcase_ASCII-2     1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/lowercase_unicode-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/mixedcase_unicode-2   1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/multibyte_unicode-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                           ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

The ASCII fastpath seems to be working as intended but there's still a severe regression for the lowercase unicode testcase.