@bt90 2024-01-25
Doing things in single pass:

```
goos: linux
goarch: amd64
pkg: github.com/syncthing/syncthing/lib/fs
cpu: AMD EPYC 7763 64-Core Processor                
                                     │ /tmp/old.txt │            /tmp/new.txt             │
                                     │    sec/op    │   sec/op     vs base                │
UnicodeLowercase/ASCII_lowercase-2      47.80n ± 5%   17.77n ± 9%  -62.83% (p=0.000 n=10)
UnicodeLowercase/ASCII_mixedcase-2     193.65n ± 9%   53.66n ± 3%  -72.29% (p=0.000 n=10)
UnicodeLowercase/Unicode_lowercase-2    124.3n ± 2%   161.6n ± 7%  +29.96% (p=0.000 n=10)
UnicodeLowercase/Unicode_mixedcase-2    345.9n ± 9%   292.4n ± 7%  -15.48% (p=0.000 n=10)
UnicodeLowercase/Unicode_multibyte-2    538.5n ± 2%   564.7n ± 5%   +4.86% (p=0.000 n=10)
geomean                                 184.6n        120.5n       -34.71%

                                     │ /tmp/old.txt │            /tmp/new.txt             │
                                     │     B/op     │    B/op     vs base                 │
UnicodeLowercase/ASCII_lowercase-2     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/ASCII_mixedcase-2     24.00 ± 0%     24.00 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_lowercase-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_mixedcase-2   32.00 ± 0%     32.00 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_multibyte-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                           ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                     │ /tmp/old.txt │            /tmp/new.txt             │
                                     │  allocs/op   │ allocs/op   vs base                 │
UnicodeLowercase/ASCII_lowercase-2     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/ASCII_mixedcase-2     1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_lowercase-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_mixedcase-2   1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_multibyte-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                           ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```