@bt90 2024-01-28
Latest numbers:

```
goos: windows
goarch: amd64
pkg: github.com/syncthing/syncthing/lib/fs
cpu: Intel(R) Core(TM) i7-6700K CPU @ 4.00GHz
                                     │   old.txt    │               new.txt               │
                                     │    sec/op    │   sec/op     vs base                │
UnicodeLowercase/ASCII_lowercase-8      36.62n ± 1%   12.91n ± 1%  -64.75% (p=0.000 n=10)
UnicodeLowercase/ASCII_mixedcase-8     181.95n ± 1%   63.24n ± 1%  -65.25% (p=0.000 n=10)
UnicodeLowercase/Unicode_lowercase-8    111.2n ± 1%   125.1n ± 1%  +12.51% (p=0.000 n=10)
UnicodeLowercase/Unicode_mixedcase-8    290.2n ± 0%   188.7n ± 1%  -34.96% (p=0.000 n=10)
UnicodeLowercase/Unicode_multibyte-8    479.7n ± 0%   488.5n ± 0%   +1.85% (p=0.000 n=10)
geomean                                 159.5n        98.80n       -38.04%

                                     │   old.txt    │               new.txt               │
                                     │     B/op     │    B/op     vs base                 │
UnicodeLowercase/ASCII_lowercase-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/ASCII_mixedcase-8     24.00 ± 0%     24.00 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_lowercase-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_mixedcase-8   32.00 ± 0%     32.00 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_multibyte-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                           ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                     │   old.txt    │               new.txt               │
                                     │  allocs/op   │ allocs/op   vs base                 │
UnicodeLowercase/ASCII_lowercase-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/ASCII_mixedcase-8     1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_lowercase-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_mixedcase-8   1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
UnicodeLowercase/Unicode_multibyte-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                           ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```